Simplify project imports/exports
