+++ tsconfig.json
@@ -1,12 +1,13 @@
 {
   "compilerOptions": {
-    "module": "CommonJS",
+    "allowJs": true,
+    "module": "commonjs",
     "outDir": "./dist",
     "rootDir": "./src",
     "target": "ES2015",
     "noImplicitAny": true,
     "lib": ["ES6", "dom"],
-    "declaration": true,
+    // "declaration": true,
     "strict": true
   },
   "include": ["src/**/*.ts"],
